+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
 ## [Unreleased]
 ### Added
 - Location multiplier configs. Allows difficulty to be increased/decreased in different dimensions and biomes.
+- Lunar cycle configs are back
 
 ## [2.0.3] - 2019-04-12
 ### Added